+++ packages/ug-layout-angular/src/AngularViewFactory.ts
@@ -147,13 +147,16 @@ export class AngularViewFactory extends ViewFactory {
       injector
     );
 
-    this._setupInputs<T>(componentFactory, componentRef);
+    const unbindInputs = this._setupInputs<T>(componentFactory, componentRef);
 
     componentRef.instance[COMPONENT_REF_KEY] = componentRef;
     viewContainer.mount(componentRef.location.nativeElement);
 
     viewContainer.destroyed
-      .subscribe(() => this._onComponentDestroy(componentRef));
+      .subscribe(() => {
+        unbindInputs();
+        this._onComponentDestroy(componentRef, componentFactory);
+      });
 
     viewContainer.attached
       .subscribe(() => this._onAttachChange(true, componentRef, viewContainer));
@@ -205,7 +208,9 @@ export class AngularViewFactory extends ViewFactory {
     }
   }
 
-  private _setupInputs<T>(factory: ComponentFactory<T>, componentRef: ComponentRef<T>): void {
+  private _setupInputs<T>(factory: ComponentFactory<T>, componentRef: ComponentRef<T>): () => void {
+    let isBound: boolean = true;
+
     for (const input of factory.inputs) {
       const descriptor = Object.getOwnPropertyDescriptor(componentRef.instance, input.propName)
         // Account for getter/setters
@@ -228,11 +233,15 @@ export class AngularViewFactory extends ViewFactory {
               set.call(this, value);
             }
 
-            componentRef.changeDetectorRef.detectChanges();
+            if (isBound) {
+              componentRef.changeDetectorRef.detectChanges();
+            }
           }
         });
       }
     }
+
+    return () => isBound = false;
   }
 
   private _onDestroyNotify<T>(componentRef: ComponentRef<T>, viewContainer: ViewContainer<T>): void {
@@ -261,7 +270,7 @@ export class AngularViewFactory extends ViewFactory {
     }
   }
 
-  private _onComponentDestroy<T>(componentRef: ComponentRef<T>): void {
+  private _onComponentDestroy<T>(componentRef: ComponentRef<T>, componentFactory: ComponentFactory<T>): void {
     const index = this._viewContainerRef.indexOf(componentRef.hostView);
 
     if (index !== -1) {
@@ -269,6 +278,14 @@ export class AngularViewFactory extends ViewFactory {
     }
 
     componentRef.destroy();
+
+    for (const output of componentFactory.outputs) {
+      if (typeof componentRef.instance === 'object'
+        && typeof componentRef.instance[output.propName] === 'object'
+        && typeof componentRef.instance[output.propName].complete === 'function') {
+        componentRef.instance[output.propName].complete();
+      }
+    }
   }
 
   protected getNg1Providers(providers: { [key: string]: any }, viewContainer: ViewContainer<any>): { [key: string]: any } {