+++ packages/ug-layout-angular/src/AngularViewTestability.ts
@@ -5,51 +5,121 @@ import {
   forkJoin
 } from 'rxjs';
 import { takeUntil, first } from 'rxjs/operators';
-import { RootLayout, ViewManager, ViewContainerStatus } from 'ug-layout';
+import { RootLayout, ViewManager, ViewContainerStatus, ViewContainer } from 'ug-layout';
 
+/**
+ * Class used for protractor testing.
+ * @export
+ * @class AngularViewTestability
+ */
 export class AngularViewTestability {
   private _layout: RootLayout | null = null;
   private _isStable: boolean = true;
-  private _checking = new Subject();
+  private _layoutSource = new Subject<RootLayout | null>();
+  private _pendingUpdate = new Subject<ViewContainer<any>[]>();
+  private _destroyed = new Subject();
+  private _pending: ViewContainer<any>[] = [];
+  private _clearSubscriptions = Observable.merge(
+    this._destroyed.asObservable(),
+    this._layoutSource.asObservable());
 
   constructor(layout?: RootLayout | null) {
+    this._layoutSource.subscribe(layoutOrNull => this.layoutChanged(layoutOrNull));
+
     if (layout) {
       this.setLayout(layout);
     }
   }
 
   setLayout(layout: RootLayout | null): void {
-    this._layout = layout;
+    this._layoutSource.next(layout);
   }
 
   isStable(): boolean {
-    return this._isStable;
+    return this._pending.length === 0;
   }
 
   whenStable(callback: Function): void {
-    this._isStable = false;
-    this._checking.next();
-
-    if (!this._layout) {
+    if (this.isStable()) {
       return callback();
     }
 
-    const viewManager = this._layout.injector.get(ViewManager);
-
-    forkJoin(
-      ...[ ...viewManager.entries() ]
-        .filter(view => view.isVisible)
-        .map(view => view.status
-          .pipe(first(status => status === ViewContainerStatus.READY, ViewContainerStatus.READY)))
-    )
-      .pipe(takeUntil(this._checking))
-      .subscribe(() => {
-        this._isStable = true;
-        callback();
-      });
+    // this._pendingUpdate
+    //   .first(pending => pending.length === 0, undefined, true)
+    //   .subscribe(() => callback());
   }
 
   findProviders(using: any, provider: string, exactMatch: boolean): any[] {
     return [];
+  }
+
+  destroy(): void {
+    this._destroyed.next();
+    this._layoutSource.complete();
+    this._pendingUpdate.complete();
+    this._destroyed.complete();
+  }
+
+  private remove(container: ViewContainer<any>): void {
+    const index = this._pending.indexOf(container);
+
+    if (index !== -1) {
+      this._pending.splice(index, 1);
+      this._pendingUpdate.next(this._pending);
+    }
+  }
+
+  private add(container: ViewContainer<any>): void {
+    if (this._pending.indexOf(container) === -1) {
+      this._pending.push(container);
+      this._pendingUpdate.next(this._pending);
+    }
+  }
+
+  private clear(): void {
+    this._pending = [];
+    this._pendingUpdate.next(this._pending);
+  }
+
+  private layoutChanged(layoutOrNull: RootLayout | null): void {
+    this.clear();
+
+    if (!layoutOrNull) {
+      return;
+    }
+
+
+    const viewManager = layoutOrNull.injector.get(ViewManager);
+
+    for (const container of viewManager.entries()) {
+      this.subscribeToContainerEvents(container);
+    }
+
+    viewManager.registered
+      .takeUntil(this._clearSubscriptions)
+      .subscribe(({ container }) => this.subscribeToContainerEvents(container));
+
+    viewManager.unregistered
+      .takeUntil(this._clearSubscriptions)
+      .subscribe(({ container }) => this.remove(container));
+  }
+
+  private subscribeToContainerEvents(container: ViewContainer<any>): void {
+    Observable.combineLatest(
+      container.visibilityChanges.distinctUntilChanged(),
+      container.componentInitialized.distinctUntilChanged(),
+      // Emit initially so combine latest doesn't wait for attach events.
+      Observable.merge(container.attached, container.detached)
+        .startWith(true)
+        .distinctUntilChanged()
+    )
+      .takeUntil(this._clearSubscriptions)
+      .subscribe(([ isVisible, isInitialized ]) => {
+        if (isVisible && !isInitialized && container.isAttached) {
+          this.add(container);
+        } else {
+          this.remove(container);
+        }
+      });
   }
 }