+++ packages/ug-layout/src/stack/index.ts
@@ -7,3 +7,4 @@ export * from './StackItemContainer';
 export * from './StackTab';
 export * from './TabCloseEvent';
 export * from './StackSerializer';
+export * from './TabSelectionEvent';